add: save forecast in database

+++ app/src/main/java/com/example/weather/room/Forecast.kt
@@ -10,5 +10,5 @@ data class Forecast(
 
     var city: String,
 
-    var time: Long
+    var time: Int
 )
